simplify plus bug fixes
