+++ tpot2/builtin_modules/imputer.py
@@ -1,0 +1,137 @@
+# %%
+# -*- coding: utf-8 -*-
+
+"""Copyright (c) 2015 The auto-sklearn developers. All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+
+  a. Redistributions of source code must retain the above copyright notice,
+     this list of conditions and the following disclaimer.
+  b. Redistributions in binary form must reproduce the above copyright
+     notice, this list of conditions and the following disclaimer in the
+     documentation and/or other materials provided with the distribution.
+  c. Neither the name of the auto-sklearn Developers  nor the names of
+     its contributors may be used to endorse or promote products
+     derived from this software without specific prior written
+     permission.
+
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ARE DISCLAIMED. IN NO EVENT SHALL THE REGENTS OR CONTRIBUTORS BE LIABLE FOR
+ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY
+OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH
+DAMAGE.
+"""
+#TODO support np arrays
+
+import numpy as np
+from scipy import sparse
+
+from sklearn.base import BaseEstimator, TransformerMixin
+from sklearn.utils import check_array
+from sklearn.preprocessing import OneHotEncoder
+import sklearn
+import sklearn.impute
+
+import pandas as pd
+from pandas.api.types import is_numeric_dtype
+import sklearn.compose
+
+
+class ColumnSimpleImputer(BaseEstimator, TransformerMixin):
+    def __init__(self,  columns="all",         
+                        missing_values=np.nan,
+                        strategy="mean",
+                        fill_value=None,
+                        copy=True,
+                        add_indicator=False,
+                        keep_empty_features=False,):
+        
+        self.columns = columns
+        self.missing_values = missing_values
+        self.strategy = strategy
+        self.fill_value = fill_value
+        self.copy = copy
+        self.add_indicator = add_indicator
+        self.keep_empty_features = keep_empty_features
+
+
+    def fit(self, X, y=None):
+        """Fit OneHotEncoder to X, then transform X.
+
+        Equivalent to self.fit(X).transform(X), but more convenient and more
+        efficient. See fit for the parameters, transform for the return value.
+
+        Parameters
+        ----------
+        X : array-like or sparse matrix, shape=(n_samples, n_features)
+            Dense array or sparse matrix.
+        y: array-like {n_samples,} (Optional, ignored)
+            Feature labels
+        """
+
+        if self.columns == "categorical" or self.columns == "numeric" and not isinstance(X, pd.DataFrame):
+            raise ValueError(f"Invalid value for columns: {self.columns}. "
+                             "Only 'all' or <list> is supported for np arrays")
+
+        if self.columns == "categorical":
+            self.columns_ = list(X.select_dtypes(exclude='number').columns)
+        elif self.columns == "numeric":
+            self.columns_ =  [col for col in X.columns if is_numeric_dtype(X[col])]
+        elif self.columns == "all":
+            if isinstance(X, pd.DataFrame):
+                self.columns_ = X.columns
+            else:
+                self.columns_ = list(range(X.shape[1]))
+        elif isinstance(self.columns, list):
+            self.columns_ = self.columns
+        else:
+            raise ValueError(f"Invalid value for columns: {self.columns}")
+        
+        self.imputer = sklearn.impute.SimpleImputer(missing_values=self.missing_values,
+                                                    strategy=self.strategy,
+                                                    fill_value=self.fill_value,
+                                                    copy=self.copy,
+                                                    add_indicator=self.add_indicator,
+                                                    keep_empty_features=self.keep_empty_features)
+        
+        if isinstance(X, pd.DataFrame):
+            self.imputer.set_output(transform="pandas")
+
+        if isinstance(X, pd.DataFrame):
+            self.imputer.fit(X[self.columns_], y)
+        else:
+            self.imputer.fit(X[:, self.columns_], y)
+
+        return self
+
+    def transform(self, X):
+        """Transform X using one-hot encoding.
+
+        Parameters
+        ----------
+        X : array-like or sparse matrix, shape=(n_samples, n_features)
+            Dense array or sparse matrix.
+
+        Returns
+        -------
+        X_out : sparse matrix if sparse=True else a 2-d array, dtype=int
+            Transformed input.
+        """
+        if isinstance(X, pd.DataFrame):
+            X = X.copy()
+            X[self.columns_] = self.imputer.transform(X[self.columns_])
+            return X
+        else:
+            X = np.copy(X)
+            X[:, self.columns_] = self.imputer.transform(X[:, self.columns_])
+            return X
+
+
